+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# HEAD
+
+Bug fix:
+- Ensure constraints and properties are string-keyed so that they match regardless of which are used (#33)
+
 # v0.10.0 (2017-09-15)
 
 Feature: